+++ src/App.tsx
@@ -1,30 +1,11 @@
 import { RouterProvider } from 'react-router-dom';
 import { QueryClient, QueryClientProvider, QueryErrorResetBoundary } from '@tanstack/react-query';
 import router from './router/Router';
-import ReactDOM from 'react-dom/client';
 import { ErrorBoundary } from 'react-error-boundary';
 import FallbackUI from './pages/FallbackUI/FallbackUI';
-import React, { Suspense } from 'react';
+import { Suspense } from 'react';
 import Loading from './pages/Loading/Loading';
 import GlobalStyle from './style/GlobalStyle';
-
-async function enableMocking() {
-  if (process.env.NODE_ENV !== 'development') {
-    return;
-  }
-
-  const { worker } = await import('./mocks/browser');
-  return worker.start();
-}
-
-enableMocking().then(() => {
-  const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
-  root.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>,
-  );
-});
 
 function App() {
   const queryClient = new QueryClient();
